add `.pylintrc` to quiet down `pylint`
this is really annoying. `pylint` does not care about `setup.cfg`, and
there seems to be no way to make it read a different configuration file
other than through command line arguments. but it is so hard to
configure `pyls` in `nvim`, and then also `nvim-lsp` is broken in
`nixpkgs`.