structure types more strictly, with less redundancy
the next step from here is to perform input validation in constructors.
the core logic should prevent creating invalid objects (semantically,
but also syntactically when represented as a string) altogether.

when an object is created, only part of the semantics can be checked.
like when creating a new message, it must have a non-empty label which
adheres to some syntactic rules, but until we try to place it in the
document, we cannot know if there is a name collision in the context we
try to place it into. this higher-level constraint must be either enforced at
a higher level, or the context (such as the object's parent) needs to be
passed to the constructor.

an open question is still how to easily implement a representation of
this kind of interaction. since it is supposed to be forbidden to create
half-baked, invalid objects, those cannot be simply inserted into the
document at the cursor position and then used as part of the presentation
just as any other object. there probably needs to be a scaffolding
widget, which only really creates the object once it passes validation.
until then it is just a temporary placeholder with its own presentation
logic.