validate only on setters and insertion
after some consideration I have decided to separate instantiation of
objects and checking that they are modified into valid states. this is
how I expect the UI logic to interact with this API. it would not be
meaningful to constrain creation of objects by validating their required
parameters, as there is no obvious mechanism to tell the client which
parameter was invalid. instead we now create object without parameters,
but modifiying any attribute goes through mandatory validation, as well
as adding the object to the document.

the class hierarchy is still a mess and full of linter errors, also the
interface is by far incomplete, but the general idea is covered by
tests. I still have to find a way to somewhat succinctly encapsulate the
validation logic for fields in different types of `Definition`s. really
there is a `MutableSequence` of fields in every definition, and write
operations must be guarded by validation.