clean up type hierarchy
while this is a bit more tidy, I came to the conclusion that Python is
simply not the right tool for the job to create a safe interface. there
are so many ways to manipulate objects, and I would have to manually
close all common loopholes to actually achieve the goal of making it
impossible to create invalid states. a statically typed language with
a strong type system would be much more helpful here. unfortunately the
only statically typed langauge I am fluent with is Go, but its type
system is everything but strong.