partially implement validation based on parent
unfortunately due to the struct embedding setup we now have n-fold
bookkeeping for the parent. for example, types with a parent which
contain a label must set that label's parent, too, so that it can
validate itself against it. there still needs to be some cleanup to make
it more robust and consistent. but in principle it works as covered by
the tests.