give identifiers and numbers a parent
while the original idea was to enable recursive validation by excluding
the object under validation from the comparison, it turns out that by
construction this is unnecessary: invalid nested structures cannot be
inserted into collections, assuming the implementation is correct.
the remaining and substantial use case are reserved labels and field
numbers, where we *must* exclude the object under validation from
comparison with others in the collection, as it contains a collection of
those items itself.