simplify getting/setting number values internally
match for concrete types internally. there is no advantage using
exported types for that purpose.

still thinking about returning number pointers for easier distinction in
a UI client. the alternative is to force the client to validate each
field and display it accordingly, but the `uint` field will be set to
0 initially and thus display that value, which does not make much sense
most of the time. the same problem applies to labels and types as well.

a somewhat consistent thing to do is to have `Field.GetNumber() Number`,
`Field.GetLabel() Label`, etc, the value of which may be `nil`
- specifically for objects which have not yet been inserted. for objects
which are part of the document we should give the guarantee that their
properties always have non-nil, semantically valid values.

on the other hand that does not *force* the client to check for `nil` on
those values, so that contract can only be expressed in documentation,
which is too soft for my taste. have to balance ease of use and strong
typing...