write out complete interface, add motivating documentation
the real driver here was that `Message` both contains `Definition`s and
`MessageField`s, which both require `validateLabel()` on their parent to
be present - but Go does not allow overlap on method names, even for the
same signature (which is reasonable, though). so it would mean either to
abandon making the interfaces `Message` implements explicit as a special
case - which I think is a bit confusing, especially given the central
role of that type - or dropping that convention entirely, which happened
with this change. also it seems to me that overusing interface embedding
is not only non-idiomatic Go, it is also makes such a large construction
really hard to follow.

I used the opportunity to define the complete outward interface of the
package, even if strictly speaking it is way over the top. there is no
need to have `Document` as an interface, as we can easily start from an
empty type, so this is just for completeness such that the file is
self-sufficient. also having most unexported methods is superfluous, but
my motivation here was to have some compiler assistance for implementing
the concrete types. maybe those will go away eventually and be replaced
by internal interfaces which only cover the behavior as required case by
case.