implement `New*` interfaces (mostly)
this change got out of hand, I have lost track. it is probably full of
errors. Go's lack of generics is disconcerting, now we have a mess of
copypasta.
the most notorious change is that checking `FieldNumber`s for
`nil` must happen in `validateAs*`, since there we have a concrete type
and would pass it on as an interface, which will not be nil at the
receiver unless explicitly set as such by the caller.