extract core logic
struggling to find a good balance between capturing semantic and
representational aspects, finally I arrived at the insight that they are
seaprate concerns. now they will be handled separately. this simplifies
matters a lot!

the core logic does not hold presentation information such as field
order. the package managing field order will wrap the core logic and its
types entirely. not sure yet if this should happen in a separate
(enclosing) package or in the same one, where the types have some
variant of the names, such as `Message_`.